requirements.txt
openai>=1.0.0
anthropic>=0.25.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
requests>=2.31.0
python-dotenv>=1.0.0
